ui 3

+++ src/main/java/model/Entity.java
@@ -22,7 +22,7 @@ public class Entity {
 	public Owner owner;
 	@SerializedName("velocity")
 	@Expose
-	public Integer velocity;
+	public Double velocity;
 	@SerializedName("angle")
 	@Expose
 	public Double angle;
